flow: split grove/lcd* nodes into their own module.
This way we liberate the other aio-dependant nodes to be built when i2c
is lacking and vice-versa. Aliases where made to the original names.
