+++ src/lib/flow/Kconfig
@@ -69,6 +69,7 @@ source "src/modules/flow/iio/Kconfig"
 source "src/modules/flow/int/Kconfig"
 source "src/modules/flow/json/Kconfig"
 source "src/modules/flow/keyboard/Kconfig"
+source "src/modules/flow/jhd1313m1/Kconfig"
 source "src/modules/flow/led-7seg/Kconfig"
 source "src/modules/flow/led-strip/Kconfig"
 source "src/modules/flow/location/Kconfig"
