+++ src/modules/flow/grove/Kconfig
@@ -1,4 +1,4 @@
 config FLOW_NODE_TYPE_GROVE
 	tristate "Node type: grove"
-	depends on USE_I2C && FLOW_NODE_TYPE_AIO
+	depends on FLOW_NODE_TYPE_AIO
 	default m
