+++ src/modules/flow/jhd1313m1/Kconfig
@@ -1,0 +1,4 @@
+config FLOW_NODE_TYPE_JHD1313M1
+	tristate "Node type: jhd1313m1"
+	depends on USE_I2C
+	default m
