+++ src/modules/flow/jhd1313m1/Makefile
@@ -1,0 +1,3 @@
+obj-$(FLOW_NODE_TYPE_JHD1313M1) += jhd1313m1.mod
+obj-jhd1313m1-$(FLOW_NODE_TYPE_JHD1313M1) := jhd1313m1.json jhd1313m1.o
+obj-jhd1313m1-$(FLOW_NODE_TYPE_JHD1313M1)-type := flow
